Decouple the area intersection filter
